+++ components/account/SignIn.js
@@ -1,80 +1,126 @@
-import { View, Text, StyleSheet,TextInput,Pressable, TouchableOpacity, SafeAreaView, DevSettings } from 'react-native'
+import { View, Text, StyleSheet,TextInput,Pressable, TouchableOpacity, SafeAreaView } from 'react-native'
 import React, { useState } from 'react'
-import { background, format, submit_button, heading, submit_button_text } from "../Features/Design.js"
 
 export default function Signin({navigation})
 {
-    const [username, setUsername] = useState('username');
+  try{  
+  const [username, setusername] = useState('username');
     const [password,setPassword] = useState('password');
 
+
+    const confirmLogin = async ()=> {
+     const userData = {
+      username: username,
+      password: password,
+     };
+
+     const response = await fetch('http://127.0.0.1:3000/userLogin',{
+        method: 'POST',
+        headers:{
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(userData),
+     });
+
+     if(response.ok){
+      //user signed in successfully
+      alert.alert('signed in Successfully');
+      navigation.navigate("Home Page");
+     }
+
+     else{
+      alert.alert("Account not found")
+     }
+    }
+  }catch(error){
+    Alert.alert('Error','Something went wrong. Please try again');
+  }
+    
     return(
-        <View style = {background}>
-          <View style={styles.stripe}>
-          </View>
-            <Text style = {heading}>Username:</Text>
+        <View style = {styles.container}>
+            <Text style = {styles.boldText}>Email:</Text>
             <TextInput 
             style = {styles.input}
-            placeholder='enter username...'
-            onChangeText={(text) => setUsername(text)}
+            placeholder='Enter Email...'
+            onChangeText={(text) => setEmail(text)}
             />
 
-          <Text style = {heading}>Password:</Text>
+          <Text style = {styles.boldText}>Password:</Text>
             <TextInput 
             style = {styles.input}
-            placeholder='enter password...'
+            placeholder='Enter Password...'
             secureTextEntry={true}
             onChangeText={(text) => setPassword(text)}
             />
-            
-            <TouchableOpacity 
-            style = {submit_button}
-            onPress={() => navigation.navigate("Home Page")}>
-              <View style={format}>
-                <Text style={submit_button_text}>
-                  SUBMIT
-                </Text>
-              </View>
+            <View>
+            <TouchableOpacity style = {styles.button}>
+              <Text
+                style={styles.text}
+              >
+                SUBMIT
+              </Text>
             </TouchableOpacity>
-            
+            </View>
             <TouchableOpacity>
                 <Text style={styles.forgot}>
                 Forgot Username/Password
                 </Text>
-            </TouchableOpacity>
-
+              </TouchableOpacity>
             <TouchableOpacity
             onPress={() => navigation.navigate("Create Account")}>
               <Text>create account</Text> 
             </TouchableOpacity>
-            
             <TouchableOpacity
             onPress={() => navigation.navigate("Main Menu")}>
-              <Text>continue as guest</Text>            
+              <Text>continue as guest</Text>
+              
             </TouchableOpacity>
         </View>
     )
 }
 
 const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: '#afd8c9',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
     input:{
         borderWidth:1,
         borderColor: '#777',
         padding:8,
-        margin:'3%',
-        width:'50%',
+        margin:10,
+        width:200,
         backgroundColor: '#fff',
         borderRadius: 50
-    }, 
+    },
+    boldText:{
+      fontWeight: 'bold',
+      fontSize: 18
+    },
+    button:{
+        marginTop: 20,
+        borderWidth: 1,
+        height: 40,
+        width: 150,
+        borderColor: '#777',
+        backgroundColor: '#efca66',
+        borderRadius: 10
+    },
+    text: {
+      justifyContent: 'center',
+      alignSelf: 'center',
+      marginTop: 5,
+      fontSize: 20,
+      marginBottom: 5
+    },
     forgot: {
       marginTop: 5,
       color: 'gray',
       fontSize: 12
-    },
-    stripe: {
-      backgroundColor: '#9ab880',
-      width: '100%',
-      height: '10%',
-      marginBottom: 40,
-      marginTop: '35%'
-    },       
+    }      
 });
+    
+
+
